// src/services/types/value-object.enum.ts
export enum AccountStatus {
  INACTIVE = "InActive",
  ACTIVE = "Active",
  BANNED = "Banned",
}

export enum Gender {
  FEMALE = "Female",
  MALE = "Male",
}

export enum PostStatus {
  PENDING = "Pending",
  APPROVED = "Approved",
  REJECTED = "Rejected",
  HIDDEN = "Hidden",
}

export enum ActionType {
  APPROVED = "Approved",
  REJECTED = "Rejected",
  SUSPENDED = "Suspended",
}

export enum InteriorCondition {
  FULL = "Full",
  NONE = "None",
}

export enum AppointmentStatus {
  PENDING = "Pending",
  REJECT = "Reject",
  ACCEPT = "Accept",
}

export enum MultimediaType {
  VIDEO = "Video",
  IMAGE = "Image",
}

export enum EntityParticipantType {
  ADMIN = "Admin",
  CUSTOMER = "Customer",
}